Added lifecycle hanlder in separate file to handle diff version
